Add files via upload
Time, when all used to use keypad phone, this was one of the most time passing game if I am not wrong. Nevertheless I used to play that game a lot. Being a kid was actually unknown about the working of the game, but today find it pleased to complete the coding of the same game which I used to play for hours.